Move NewRequestLogger from cmd to api package
