+++ cmd/cmd.go
@@ -1,13 +1,9 @@
 package cmd
 
 import (
-	"net/http"
 	"os"
 	"strings"
-	"time"
 
-	"github.com/gitpods/gitpods/pkg/api"
-	"github.com/go-chi/chi/middleware"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 )
@@ -70,25 +66,4 @@ func NewLogger(json bool, loglevel string) log.Logger {
 		"ts", log.DefaultTimestampUTC,
 		"caller", log.DefaultCaller,
 	)
-}
-
-func NewRequestLogger(logger log.Logger) func(next http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			start := time.Now()
-
-			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
-			next.ServeHTTP(ww, r)
-
-			level.Debug(logger).Log(
-				"request", api.GetRequestID(r.Context()),
-				"proto", r.Proto,
-				"method", r.Method,
-				"status", ww.Status(),
-				"path", r.URL.Path,
-				"duration", time.Since(start),
-				"bytes", ww.BytesWritten(),
-			)
-		})
-	}
 }
