+++ pkg/api/request_id.go
@@ -3,7 +3,11 @@ package api
 import (
 	"context"
 	"net/http"
+	"time"
 
+	"github.com/go-chi/chi/middleware"
+	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
 	"github.com/satori/go.uuid"
 )
 
@@ -28,4 +32,25 @@ func GetRequestID(ctx context.Context) string {
 		return reqID
 	}
 	return ""
+}
+
+func NewRequestLogger(logger log.Logger) func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			start := time.Now()
+
+			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
+			next.ServeHTTP(ww, r)
+
+			level.Debug(logger).Log(
+				"request", GetRequestID(r.Context()),
+				"proto", r.Proto,
+				"method", r.Method,
+				"status", ww.Status(),
+				"path", r.URL.Path,
+				"duration", time.Since(start),
+				"bytes", ww.BytesWritten(),
+			)
+		})
+	}
 }
